Fixed livecd.sh to run scripts properly in chroot context when this is needed. Added livecd-checkexit.sh subscript to have better code reuse. Improved and fixed code here and there. Unfixed issues remain (mostly my own problems with sh:-)
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@2246 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ LiveCD/Subscripts/live-cleanup.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
-#
 # --cleanup
-#
+
+echo
+echo "Entering Cleanup stage..."
+
+touch /tmp/bla
+return
 
 apt-get clean
 rm -rf /tmp/*
@@ -37,9 +41,10 @@ rm -r /home/$ETOILE_USER_NAME/GNUstep/Library/Bookmark
 
 # Build cleanup
 
-rm -r /build
-apt-get -y uninstall subversion
+if [ $REMOVE_BUILD_FILES = yes ]; then
+	rm -r /build
+	apt-get -y uninstall subversion
+fi
 
 exit
 
-$SUBSCRIPT_DIR/livecd-shrink.sh
